requirements.txt
fastapi
uvicorn[standard]
jinja2
authlib
pydantic-settings
itsdangerous
httpx
motor
databricks-sdk